refactor: collation gets one owner with the bounded collator cache
The API kept an unbounded per-locale map while the web client evicted; @stll/collation owns the bounded LRU version plus the codepoint comparator. The require-cached-collator rule now scopes the package and allowlists the module that builds the collator.
