+++ apps/api/src/handlers/entities/zip-archive.ts
@@ -1,7 +1,7 @@
 // Pure helpers for streaming a folder subtree as a ZIP archive.
 // `download-zip.ts` wires these to the database, S3, and `client-zip`.
 
-import { compareCodepoint } from "@/api/lib/collation";
+import { compareCodepoint } from "@stll/collation";
 import { sanitizeFilename } from "@/api/lib/sanitize-filename";
 
 /** A folder descendant, with the fields needed to rebuild the tree. */
