+++ apps/api/src/handlers/reports/spec/render-report-spec.ts
@@ -53,7 +53,7 @@ import type {
   ReportSpec,
 } from "@/api/handlers/reports/spec/report-spec";
 import { interpolate } from "@/api/handlers/reports/spec/report-spec";
-import { compareByLocale } from "@/api/lib/collation";
+import { compareByLocale } from "@stll/collation";
 import type { AiFieldGenerator } from "@/api/lib/docx/resolve-ai-fields";
 import { ConfigurationError } from "@/api/lib/errors/tagged-errors";
 import type { PositionSeverity } from "@/api/lib/workflow/playbook-position-facets";
