+++ apps/api/src/handlers/template-packs/catalogue.ts
@@ -16,7 +16,7 @@ import type {
 } from "@stll/template-packs/schema";
 
 import { env } from "@/api/env";
-import { compareByLocale, compareCodepoint } from "@/api/lib/collation";
+import { compareByLocale, compareCodepoint } from "@stll/collation";
 import type { MemberRole } from "@/api/lib/member-roles";
 
 let catalogue: TemplatePackCatalogue | null = null;
