+++ apps/api/src/handlers/workspaces/infosoud-courts.ts
@@ -9,7 +9,7 @@ import {
 
 import { createSafeHandler } from "@/api/lib/api-handlers";
 import type { HandlerConfig } from "@/api/lib/api-handlers";
-import { compareByLocale } from "@/api/lib/collation";
+import { compareByLocale } from "@stll/collation";
 import { HandlerError } from "@/api/lib/errors/tagged-errors";
 
 import { getInfoSoudClient } from "./infosoud-common";
