+++ apps/api/src/lib/docx/discover-template.ts
@@ -14,7 +14,7 @@ import * as slimdom from "slimdom";
 
 import { parseCondition, type ConditionNode } from "@stll/template-conditions";
 
-import { compareCodepoint } from "@/api/lib/collation";
+import { compareCodepoint } from "@stll/collation";
 
 import { parseBlockTree, scanBlockDirectives } from "./block-directives";
 import { PLACEHOLDER_RE } from "./discover-placeholders";
