+++ apps/api/src/lib/entity-filters.ts
@@ -21,7 +21,7 @@ import {
 import { user } from "@/api/db/auth-schema";
 import { entities, entityVersions, fields, properties } from "@/api/db/schema";
 import type { EntityKind, FieldContent } from "@/api/db/schema-validators";
-import { compareByLocale } from "@/api/lib/collation";
+import { compareByLocale } from "@stll/collation";
 import { typedPgArray } from "@/api/lib/search/sql";
 
 // -- Types --
