+++ apps/web/package.json
@@ -58,6 +58,7 @@
     "@stll/catalogue": "workspace:*",
     "@stll/chat": "workspace:*",
     "@stll/chat-limits": "workspace:*",
+    "@stll/collation": "workspace:*",
     "@stll/conditions": "workspace:*",
     "@stll/country-codes": "workspace:*",
     "@stll/docx-core": "catalog:",
