+++ apps/web/src/components/workspaces/create-matter-dialog.logic.ts
@@ -1,4 +1,4 @@
-import { compareByLocale } from "@/lib/collation";
+import { compareByLocale } from "@stll/collation";
 
 type ExistingWorkspace = {
   client?: { id: string } | null;
