+++ apps/web/src/routes/_protected.knowledge/-components/catalogue/catalogue-browser.tsx
@@ -42,7 +42,7 @@ import {
 } from "@/components/responsive-action-toolbar";
 import { useLocale } from "@/i18n/formatting-context";
 import type { TranslationKey } from "@/i18n/types";
-import { compareByLocale } from "@/lib/collation";
+import { compareByLocale } from "@stll/collation";
 import { detached } from "@/lib/detached";
 import { userErrorFromThrown } from "@/lib/errors/user-safe";
 import type { PracticeJurisdiction } from "@/lib/jurisdictions";
