+++ apps/web/src/routes/_protected.workspaces/-filters/client-filter-popover.tsx
@@ -10,7 +10,7 @@ import { Separator } from "@stll/ui/separator";
 import { cn } from "@stll/ui/utils";
 
 import { useLocale } from "@/i18n/formatting-context";
-import { compareByLocale } from "@/lib/collation";
+import { compareByLocale } from "@stll/collation";
 import type { Workspace } from "@/lib/workspaces/types";
 
 type ClientFilterPopoverProps = {
