+++ apps/web/src/routes/_protected.workspaces/-filters/team-filter-popover.tsx
@@ -11,7 +11,7 @@ import { cn } from "@stll/ui/utils";
 import Tooltip from "@/components/tooltip";
 import { UserIdentity } from "@/components/user-avatar";
 import { useLocale } from "@/i18n/formatting-context";
-import { compareByLocale } from "@/lib/collation";
+import { compareByLocale } from "@stll/collation";
 import { getDisplayName } from "@/lib/get-display-name";
 import type { LeadFilter, Workspace } from "@/lib/workspaces/types";
