+++ knip.json
@@ -83,6 +83,9 @@
       "entry": ["src/index.ts", "src/**/*.test.ts"],
       "ignoreDependencies": ["@stll/api"]
     },
+    "packages/collation": {
+      "entry": ["src/index.ts", "src/**/*.test.ts"]
+    },
     "packages/legal-atlas": {
       "entry": ["src/index.ts", "src/**/*.test.ts"]
     },
