+++ packages/collation/src/collation.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from "bun:test";
 
-import { compareByLocale, getCollator } from "@/lib/collation";
+import { compareByLocale, compareCodepoint, getCollator } from "./collation";
 
 describe("getCollator", () => {
   test("caches one collator instance per locale", () => {
@@ -50,5 +50,23 @@ describe("compareByLocale", () => {
     expect(
       items.toSorted((a, b) => compare(a.name, b.name)).map((i) => i.name),
     ).toEqual(["hora", "chata", "ida"]);
+  });
+});
+
+describe("compareCodepoint", () => {
+  test("orders by codepoint, ignoring locale collation rules", () => {
+    // Under cs-CZ collation "ch" sorts after "h", but codepoint order must
+    // not follow that rule.
+    expect(["ia", "cha", "ha"].toSorted(compareCodepoint)).toEqual([
+      "cha",
+      "ha",
+      "ia",
+    ]);
+  });
+
+  test("is antisymmetric and reports equality as 0", () => {
+    expect(compareCodepoint("a", "a")).toBe(0);
+    expect(compareCodepoint("a", "b")).toBe(-1);
+    expect(compareCodepoint("b", "a")).toBe(1);
   });
 });
