+++ packages/collation/src/index.ts
@@ -1,0 +1,2 @@
+// Locale-aware collation, and the codepoint order for technical keys.
+export { compareByLocale, compareCodepoint, getCollator } from "./collation";
